requirements.txt
python-dotenv==1.0.0
streamlit==1.18.1
openai==0.27.7
pyaudio==0.2.13
websockets==11.0.3
altair<5